fix(snomed): Limit maximum group number to 32767 in relationship...
...create and update requests. The type remains `int` as it is easier to
work with in both persistence and runtime scenarios.

+++ snomed/com.b2international.snowowl.snomed.datastore/src/com/b2international/snowowl/snomed/datastore/request/SnomedRelationshipCreateRequest.java
@@ -52,10 +52,10 @@ public final class SnomedRelationshipCreateRequest extends BaseSnomedComponentCr
 	// Not @NotNull, it will be null if a destination ID is set
 	private RelationshipValue value; 
 
-	@NotNull @Min(0) @Max(Byte.MAX_VALUE)
+	@NotNull @Min(0) @Max(Short.MAX_VALUE)
 	private Integer group;
 
-	@NotNull @Min(0) @Max(Byte.MAX_VALUE)
+	@NotNull @Min(0) @Max(Short.MAX_VALUE)
 	private Integer unionGroup;
 
 	@NotNull
